R01-R11: Reorganizar arquitectura modular por bounded contexts (DDD)
