+++ .scannerwork/report-task.txt
@@ -2,5 +2,5 @@ projectKey=ProyectoFinal
 serverUrl=http://localhost:9000
 serverVersion=9.9.8.100196
 dashboardUrl=http://localhost:9000/dashboard?id=ProyectoFinal
-ceTaskId=AZ7Miwu7BiBmtlsPXu44
-ceTaskUrl=http://localhost:9000/api/ce/task?id=AZ7Miwu7BiBmtlsPXu44
+ceTaskId=AZ7XI1HIC2Bx9tf_myI1
+ceTaskUrl=http://localhost:9000/api/ce/task?id=AZ7XI1HIC2Bx9tf_myI1
