+++ src/main/java/com/batuhaniskr/product/category/domain/Category.java
@@ -1,0 +1,46 @@
+package com.batuhaniskr.product.category.domain;
+
+import com.batuhaniskr.product.product.domain.Product;
+
+import javax.persistence.*;
+import java.util.Set;
+
+@Entity
+@Table(name = "category")
+public class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column(nullable = false, name = "category_name")
+    private String categoryName;
+
+    @Column(nullable = false)
+    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    private Set<Product> productSet;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getCategoryName() {
+        return categoryName;
+    }
+
+    public void setCategoryName(String categoryName) {
+        this.categoryName = categoryName;
+    }
+
+    public Set<Product> getProductSet() {
+        return productSet;
+    }
+
+    public void setProductSet(Set<Product> productSet) {
+        this.productSet = productSet;
+    }
+}
