+++ src/main/java/com/batuhaniskr/product/category/domain/CategoryDTO.java
@@ -1,0 +1,35 @@
+package com.batuhaniskr.product.category.domain;
+
+import com.batuhaniskr.product.product.domain.ProductDTO;
+
+import java.util.Set;
+
+public class CategoryDTO {
+    private Integer id;
+    private String categoryName;
+    private Set<ProductDTO> productSet;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getCategoryName() {
+        return categoryName;
+    }
+
+    public void setCategoryName(String categoryName) {
+        this.categoryName = categoryName;
+    }
+
+    public Set<ProductDTO> getProductSet() {
+        return productSet;
+    }
+
+    public void setProductSet(Set<ProductDTO> productSet) {
+        this.productSet = productSet;
+    }
+}
