--- src/main/java/com/batuhaniskr/product/config/MyAccessDeniedHandler.java
@@ -1,33 +1,0 @@
-package com.batuhaniskr.product.config;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class MyAccessDeniedHandler implements AccessDeniedHandler{
-
-    private static Logger logger = LoggerFactory.getLogger(MyAccessDeniedHandler.class);
-
-    @Override
-    public void handle(HttpServletRequest httpServletRequest,
-                       HttpServletResponse httpServletResponse,
-                       AccessDeniedException e) throws IOException {
-
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication != null){
-            logger.info("'" +authentication.getName() + "'"
-            + " was trying to access protected url " + httpServletRequest.getRequestURI());
-        }
-        httpServletResponse.sendRedirect(httpServletRequest.getContextPath() + "/access-denied");
-    }
-}
