--- src/main/java/com/batuhaniskr/product/controller/UserController.java
@@ -1,55 +1,0 @@
-package com.batuhaniskr.product.controller;
-
-import com.batuhaniskr.product.dto.UserRegistrationDto;
-import com.batuhaniskr.product.model.User;
-import com.batuhaniskr.product.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-
-@Controller
-public class UserController {
-
-    private UserService userService;
-
-    @Autowired
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-
-    @RequestMapping("/access-denied")
-    public String accessDenied() {
-        return "error/access-denied";
-    }
-
-    @ModelAttribute("user")
-    public UserRegistrationDto userRegistrationDto() {
-        return new UserRegistrationDto();
-    }
-
-    @GetMapping(value = "/login")
-    public String showRegistrationForm(Model model) {
-        return "login";
-    }
-
-    @RequestMapping(value = "/registration", method = RequestMethod.POST)
-    public String registerUserAccount(@ModelAttribute("user") @Valid UserRegistrationDto userDto,
-                                      BindingResult result, Model model) {
-        User existing = userService.findByEmail(userDto.getEmail());
-        if (existing != null) {
-            result.rejectValue("email", null, "There is already an account registered with that email");
-        }
-
-        if (result.hasErrors()) {
-            return "login";
-        }
-
-        userService.save(userDto);
-
-        return "redirect:/login?success";
-    }
-}
