--- src/main/java/com/batuhaniskr/product/dto/ProductDTO.java
@@ -1,57 +1,0 @@
-package com.batuhaniskr.product.dto;
-
-import javax.validation.constraints.DecimalMin;
-import javax.validation.constraints.Min;
-import java.math.BigDecimal;
-
-public class ProductDTO {
-
-    private Integer id;
-    private String name;
-
-    @Min(0)
-    private Integer quantity;
-    @DecimalMin("0.0")
-    private BigDecimal price;
-    private CategoryDTO category;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(Integer quantity) {
-        this.quantity = quantity;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-
-    public CategoryDTO getCategory() {
-        return category;
-    }
-
-    public void setCategory(CategoryDTO category) {
-        this.category = category;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-}
