--- src/main/java/com/batuhaniskr/product/dto/UserRegistrationDto.java
@@ -1,75 +1,0 @@
-package com.batuhaniskr.product.dto;
-
-import com.batuhaniskr.product.validation.FieldMatch;
-import org.hibernate.validator.constraints.Email;
-import org.hibernate.validator.constraints.NotEmpty;
-
-import javax.validation.constraints.AssertTrue;
-import javax.validation.constraints.Pattern;
-
-@FieldMatch.List({
-        @FieldMatch(first = "password", second = "confirmPassword", message = "The password fields must match"),
-        @FieldMatch(first = "email", second = "confirmEmail", message = "The email fields must match")
-})
-public class UserRegistrationDto {
-
-    @NotEmpty
-    private String username;
-
-    @NotEmpty
-    @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8,}$",
-            message = "Must a password be at least 8 characters and include at least one letter, one number")
-    private String password;
-
-    @NotEmpty
-    @Pattern(regexp = "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d]{8,}$",
-            message = "Must a password be at least 8 characters and include at least one letter, one number")
-    private String confirmPassword;
-
-    @Email
-    @NotEmpty
-    private String email;
-
-    @AssertTrue
-    private Boolean terms;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getConfirmPassword() {
-        return confirmPassword;
-    }
-
-    public void setConfirmPassword(String confirmPassword) {
-        this.confirmPassword = confirmPassword;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public Boolean getTerms() {
-        return terms;
-    }
-
-    public void setTerms(Boolean terms) {
-        this.terms = terms;
-    }
-}
