+++ src/main/java/com/batuhaniskr/product/product/presentation/MainController.java
@@ -1,0 +1,110 @@
+package com.batuhaniskr.product.product.presentation;
+
+import com.batuhaniskr.product.category.domain.CategoryDTO;
+import com.batuhaniskr.product.product.domain.ProductDTO;
+import com.batuhaniskr.product.category.application.CategoryService;
+import com.batuhaniskr.product.product.application.ProductService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.NoHandlerFoundException;
+
+import javax.validation.Valid;
+import java.util.List;
+import java.util.Optional;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+@Controller
+@RequestMapping("/products")
+public class MainController {
+
+    private ProductService productService;
+    private CategoryService categoryService;
+
+    private static int currentPage = 1;
+    private static int pageSize = 5;
+    private static final Logger LOG = Logger.getLogger(MainController.class.getName());
+
+    @Autowired
+    public MainController(ProductService productService, CategoryService categoryService) {
+        this.productService = productService;
+        this.categoryService = categoryService;
+    }
+
+    @GetMapping("")
+    public String index(Model model, @RequestParam("page") Optional<Integer> page,
+                        @RequestParam("size") Optional<Integer> size,
+                        @AuthenticationPrincipal UserDetails userDetails) {
+
+
+        page.ifPresent(p -> currentPage = p);
+        size.ifPresent(s -> pageSize = s);
+
+        Pageable pageable = new PageRequest(currentPage - 1, pageSize);
+        String userEmail = userDetails.getUsername();
+        Page<ProductDTO> productPage = productService.getAllProduct(pageable, userEmail);
+
+        model.addAttribute("productPage", productPage);
+
+        int totalPages = productPage.getTotalPages();
+        if (totalPages > 0) {
+            List<Integer> pageNumbers = IntStream.rangeClosed(1, totalPages)
+                    .boxed()
+                    .collect(Collectors.toList());
+            model.addAttribute("pageNumbers", pageNumbers);
+        }
+
+        return "products";
+    }
+
+
+    @GetMapping(value = "/add")
+    public String addProduct(@Valid Model model) {
+        List<CategoryDTO> categoryList = categoryService.getAllCategory();
+        model.addAttribute("product", new ProductDTO());
+        model.addAttribute("categoryList", categoryList);
+
+        return "addproduct";
+    }
+
+    @PostMapping(value = "/save")
+    public String save(ProductDTO productDTO, @AuthenticationPrincipal UserDetails userDetails) {
+        LOG.log(Level.INFO, "/ {0}",productDTO.getName());
+        String userEmail = userDetails.getUsername();
+        productService.saveProduct(productDTO, userEmail);
+
+        return "redirect:/products";
+    }
+
+    @GetMapping(value = "/delete/{id}")
+    public String deleteProduct(@PathVariable("id") Integer id) {
+        productService.deleteProduct(id);
+
+        return "redirect:/products";
+    }
+
+    @GetMapping(value = "/edit/{id}")
+    public String editProduct(@PathVariable("id") Integer id, Model model) {
+        ProductDTO product = productService.getProductById(id);
+        List<CategoryDTO> categoryList = categoryService.getAllCategory();
+        model.addAttribute("product", product);
+        model.addAttribute("categoryList", categoryList);
+
+        return "editproduct";
+    }
+
+    @ExceptionHandler(NoHandlerFoundException.class)
+    public String handlerException() {
+        return "error/404";
+    }
+}
