--- src/main/java/com/batuhaniskr/product/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.batuhaniskr.product.repository;
-
-import com.batuhaniskr.product.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-}
