+++ src/main/java/com/batuhaniskr/product/shared/validation/FieldMatch.java
@@ -1,0 +1,31 @@
+package com.batuhaniskr.product.shared.validation;
+
+import javax.validation.Payload;
+import javax.validation.Constraint;
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target({TYPE, ANNOTATION_TYPE})
+@Retention(RUNTIME)
+@Constraint(validatedBy = FieldMatchValidator.class)
+@Documented
+public @interface FieldMatch
+{
+    String message() default "{constraints.field-match}";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+    String first();
+    String second();
+
+    @Target({TYPE, ANNOTATION_TYPE})
+    @Retention(RUNTIME)
+    @Documented
+    @interface List
+    {
+        FieldMatch[] value();
+    }
+}
