+++ src/main/java/com/batuhaniskr/product/user/infrastructure/UserRepository.java
@@ -1,0 +1,10 @@
+package com.batuhaniskr.product.user.infrastructure;
+
+import com.batuhaniskr.product.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByEmail(String email);
+}
